# sdks/dotnet/api/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: hello@zahidaramai.com
    name: KLCUBE NETWORK Enterprise
  description: |
    API middleware gateway for Malaysia's MyInvois e-invoicing system (LHDN).

    **Client:** Hashmato | **Version:** 1.1.1

    ## Features
    - 4 Submission Endpoints (Consolidate, JustSave, Buyer, Personal)
    - JWT Authentication with Access & Refresh Tokens
    - User/Role/Company Management
    - Document Operations (List, Status, PDF, Cancel)
    - Digital Signing v1.1 (XAdES signatures)
    - Background Status Polling

    ## Authentication
    All endpoints except `/api/v1/auth/login` require JWT authentication.
    Include the access token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```
  title: HashLHDN MyInvois Middleware API
  version: 1.1.1
servers:
  - description: Local Development
    url: http://localhost:3000
  - description: Production
    url: https://api.hashlhdn.com
...
paths:
...
  /api/v1/auth/login:
    post:
      description: "Authenticate with email/password, returns access and refresh tokens"
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
          description: Login successful
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Authentication required
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Rate limit exceeded
          headers:
            Retry-After:
              description: Seconds to wait before retrying
              explode: false
              schema:
                type: integer
              style: simple
      summary: Login
      tags:
        - Authentication
  /api/v1/auth/logout:
    post:
      description: Invalidate current tokens
      operationId: logout
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoutRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/logout_200_response"
          description: Logout successful
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Rate limit exceeded
          headers:
            Retry-After:
              description: Seconds to wait before retrying
              explode: false
              schema:
                type: integer
              style: simple
      security:
        - bearerAuth: []
      summary: Logout
      tags:
        - Authentication
  /api/v1/auth/refresh:
    post:
      description: Exchange refresh token for new access token
      operationId: refreshToken
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResponse"
          description: Token refreshed
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Authentication required
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Rate limit exceeded
          headers:
            Retry-After:
              description: Seconds to wait before retrying
              explode: false
              schema:
                type: integer
              style: simple
      summary: Refresh access token
      tags:
        - Authentication
  /api/v1/auth/me:
    get:
      description: Returns details of the authenticated user
      operationId: getCurrentUser
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUserResponse"
          description: Current user details
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Authentication required
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Rate limit exceeded
          headers:
            Retry-After:
              description: Seconds to wait before retrying
              explode: false
              schema:
                type: integer
              style: simple
      security:
        - bearerAuth: []
      summary: Get current user
      tags:
        - Authentication
  /api/v1/auth/switch-company:
    post:
      description: Switch to a different company context
      operationId: switchCompany
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/switchCompany_request"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
          description: Company switched
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Rate limit exceeded
          headers:
            Retry-After:
              description: Seconds to wait before retrying
              explode: false
              schema:
                type: integer
              style: simple
      security:
        - bearerAuth: []
      summary: Switch active company
      tags:
        - Authentication
...
  /api/v1/users/{id}:
    delete:
      description: Permanently delete a user from the system
      operationId: deleteUser
      parameters:
        - explode: false
          in: path
          name: id
          required: true
          schema:
            type: string
          style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/deleteUser_200_response"
          description: User deleted
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Rate limit exceeded
          headers:
            Retry-After:
              description: Seconds to wait before retrying
              explode: false
              schema:
                type: integer
              style: simple
      security:
        - bearerAuth: []
      summary: Delete user
      tags:
        - Users
    get:
      description: Retrieve user details including role and company assignments
      operationId: getUser
      parameters:
        - explode: false
          in: path
          name: id
          required: true
          schema:
            type: string
          style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
          description: User details
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Resource not found
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Rate limit exceeded
          headers:
            Retry-After:
              description: Seconds to wait before retrying
              explode: false
              schema:
                type: integer
              style: simple
      security:
        - bearerAuth: []
      summary: Get user by ID
      tags:
        - Users
    put:
      description: "Update user details such as name, email, or active status"
      operationId: updateUser
      parameters:
        - explode: false
          in: path
          name: id
          required: true
          schema:
            type: string
          style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
          description: User updated
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Rate limit exceeded
          headers:
            Retry-After:
              description: Seconds to wait before retrying
              explode: false
              schema:
                type: integer
              style: simple
      security:
        - bearerAuth: []
      summary: Update user
      tags:
        - Users
...
components:
...
  schemas:
    ErrorResponse:
      example:
        error:
          code: code
          details:
            - "{}"
            - "{}"
          message: message
      properties:
        error:
          $ref: "#/components/schemas/ErrorResponse_error"
      required:
        - error
      type: object
...
    LoginRequest:
      example:
        password: password
        email: email
      properties:
        email:
          format: email
          type: string
        password:
          format: password
          type: string
      required:
        - email
        - password
      type: object
    LoginResponse:
      example:
        expiresIn: 0
        accessToken: accessToken
        user:
          createdAt: createdAt
          companies:
            - idType: idType
              address: address
              city: city
              postalCode: postalCode
              idValue: idValue
              isActive: true
              hasCredentials: true
              createdAt: createdAt
              phone: phone
              myinvoisEnv: SANDBOX
              name: name
              tin: tin
              id: id
              state: state
              email: email
              updatedAt: updatedAt
            - idType: idType
              address: address
              city: city
              postalCode: postalCode
              idValue: idValue
              isActive: true
              hasCredentials: true
              createdAt: createdAt
              phone: phone
              myinvoisEnv: SANDBOX
              name: name
              tin: tin
              id: id
              state: state
              email: email
              updatedAt: updatedAt
          role:
            isSystem: true
            createdAt: createdAt
            permissions:
              - permissions
              - permissions
            name: name
            description: description
            id: id
            updatedAt: updatedAt
          name: name
          id: id
          isActive: true
          email: email
          updatedAt: updatedAt
        refreshToken: refreshToken
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          description: Access token expiry in seconds
          type: integer
        user:
          $ref: "#/components/schemas/User"
      type: object
    LogoutRequest:
      example:
        refreshToken: refreshToken
      properties:
        refreshToken:
          type: string
      required:
        - refreshToken
      type: object
    RefreshRequest:
      example:
        refreshToken: refreshToken
      properties:
        refreshToken:
          type: string
      required:
        - refreshToken
      type: object
    RefreshResponse:
      example:
        expiresIn: 0
        accessToken: accessToken
      properties:
        accessToken:
          type: string
        expiresIn:
          type: integer
      type: object
    CurrentUserResponse:
      example:
        companies:
          - idType: idType
            address: address
            city: city
            postalCode: postalCode
            idValue: idValue
            isActive: true
            hasCredentials: true
            createdAt: createdAt
            phone: phone
            myinvoisEnv: SANDBOX
            name: name
            tin: tin
            id: id
            state: state
            email: email
            updatedAt: updatedAt
          - idType: idType
            address: address
            city: city
            postalCode: postalCode
            idValue: idValue
            isActive: true
            hasCredentials: true
            createdAt: createdAt
            phone: phone
            myinvoisEnv: SANDBOX
            name: name
            tin: tin
            id: id
            state: state
            email: email
            updatedAt: updatedAt
        user:
          createdAt: createdAt
          companies:
            - idType: idType
              address: address
              city: city
              postalCode: postalCode
              idValue: idValue
              isActive: true
              hasCredentials: true
              createdAt: createdAt
              phone: phone
              myinvoisEnv: SANDBOX
              name: name
              tin: tin
              id: id
              state: state
              email: email
              updatedAt: updatedAt
            - idType: idType
              address: address
              city: city
              postalCode: postalCode
              idValue: idValue
              isActive: true
              hasCredentials: true
              createdAt: createdAt
              phone: phone
              myinvoisEnv: SANDBOX
              name: name
              tin: tin
              id: id
              state: state
              email: email
              updatedAt: updatedAt
          role:
            isSystem: true
            createdAt: createdAt
            permissions:
              - permissions
              - permissions
            name: name
            description: description
            id: id
            updatedAt: updatedAt
          name: name
          id: id
          isActive: true
          email: email
          updatedAt: updatedAt
      properties:
        user:
          $ref: "#/components/schemas/User"
        companies:
          items:
            $ref: "#/components/schemas/Company"
          type: array
      type: object
...
    User:
      example:
        createdAt: createdAt
        companies:
          - idType: idType
            address: address
            city: city
            postalCode: postalCode
            idValue: idValue
            isActive: true
            hasCredentials: true
            createdAt: createdAt
            phone: phone
            myinvoisEnv: SANDBOX
            name: name
            tin: tin
            id: id
            state: state
            email: email
            updatedAt: updatedAt
          - idType: idType
            address: address
            city: city
            postalCode: postalCode
            idValue: idValue
            isActive: true
            hasCredentials: true
            createdAt: createdAt
            phone: phone
            myinvoisEnv: SANDBOX
            name: name
            tin: tin
            id: id
            state: state
            email: email
            updatedAt: updatedAt
        role:
          isSystem: true
          createdAt: createdAt
          permissions:
            - permissions
            - permissions
          name: name
          description: description
          id: id
          updatedAt: updatedAt
        name: name
        id: id
        isActive: true
        email: email
        updatedAt: updatedAt
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        isActive:
          type: boolean
        role:
          $ref: "#/components/schemas/Role"
        companies:
          items:
            $ref: "#/components/schemas/Company"
          type: array
        createdAt:
          type: string
        updatedAt:
          type: string
      type: object
...
    UpdateUserRequest:
      example:
        password: password
        name: name
        isActive: true
        email: email
      properties:
        email:
          format: email
          type: string
        name:
          type: string
        isActive:
          type: boolean
        password:
          minLength: 8
          type: string
      type: object
...
    Role:
      example:
        isSystem: true
        createdAt: createdAt
        permissions:
          - permissions
          - permissions
        name: name
        description: description
        id: id
        updatedAt: updatedAt
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        permissions:
          items:
            type: string
          type: array
        isSystem:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      type: object
...
    Company:
      example:
        idType: idType
        address: address
        city: city
        postalCode: postalCode
        idValue: idValue
        isActive: true
        hasCredentials: true
        createdAt: createdAt
        phone: phone
        myinvoisEnv: SANDBOX
        name: name
        tin: tin
        id: id
        state: state
        email: email
        updatedAt: updatedAt
      properties:
        id:
          type: string
        name:
          type: string
        tin:
          type: string
        idType:
          type: string
        idValue:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        postalCode:
          type: string
        phone:
          type: string
        email:
          type: string
        isActive:
          type: boolean
        myinvoisEnv:
          enum:
            - SANDBOX
            - PROD
          type: string
        hasCredentials:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      type: object
...
    logout_200_response:
      example:
        message: Logged out successfully
      properties:
        message:
          example: Logged out successfully
          type: string
      type: object
    switchCompany_request:
      properties:
        companyId:
          type: string
      required:
        - companyId
      type: object
...
    deleteUser_200_response:
      example:
        message: message
      properties:
        message:
          type: string
      type: object
...
    ErrorResponse_error:
      example:
        code: code
        details:
          - "{}"
          - "{}"
        message: message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          items:
            type: object
          type: array
      required:
        - code
        - message
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http